normalisation params

+++ train_endec_caIIk.py
@@ -558,9 +558,9 @@ class dataset_spot(torch.utils.data.Dataset):
 
             self.max_model = np.max(self.model, axis=(1, 2, 3))
 
-            normalise_profile_params = np.zeros(30, 2)
+            normalise_profile_params = np.zeros((30, 2), dtype=np.float64)
 
-            normalise_model_params = np.zeros(450, 2)
+            normalise_model_params = np.zeros((450, 2), dtype=np.float64)
 
             normalise_profile_params[:, 0] = self.min_profile
